requirements.txt
# Data Collection
tweepy==4.14.0
beautifulsoup4==4.12.2
requests==2.31.0
yfinance==0.2.36
apache-airflow==2.8.1

# Data Processing
nltk==3.8.1
spacy==3.7.2
transformers==4.37.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
mlflow==2.9.2

# Database
psycopg2-binary==2.9.9
sqlalchemy>=1.4.28,<2.0.0
alembic==1.12.1
redis==5.0.1

# Backend
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
pytest==7.4.3
python-jwt==4.0.0

# Frontend
streamlit==1.28.2
plotly==5.18.0

# Utilities
python-dotenv==1.0.0 